Refactored `src` to be correct for gradle
