--- src/app/Main.java
@@ -1,49 +1,0 @@
-package app;
-
-import data_access.FileUserDataAccessObject;
-import entity.CommonUserFactory;
-import interface_adapters.ViewManagerModel;
-import interface_adapters.logged_in.LoggedInViewModel;
-import interface_adapters.login.LoginViewModel;
-import view.LoggedInView;
-import view.LoginView;
-import view.ViewManager;
-
-import javax.swing.*;
-import java.awt.*;
-import java.io.IOException;
-
-public class Main {
-    public static void main(String[] args) {
-
-        JFrame application = new JFrame("Login");
-        application.setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
-
-        CardLayout cardLayout = new CardLayout();
-
-        JPanel views = new JPanel(cardLayout);
-        application.add(views);
-
-        ViewManagerModel viewManagerModel = new ViewManagerModel();
-        new ViewManager(views, cardLayout, viewManagerModel);
-
-        LoginViewModel loginViewModel = new LoginViewModel();
-        LoggedInViewModel loggedInViewModel = new LoggedInViewModel();
-
-        FileUserDataAccessObject userDataAccessObject;
-        try {
-            userDataAccessObject = new FileUserDataAccessObject("./users.csv", new CommonUserFactory());
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-
-        LoginView loginView = LoginUseCaseFactory.create(viewManagerModel, loginViewModel, loggedInViewModel, userDataAccessObject);
-        views.add(loginView, loginView.viewName);
-
-        LoggedInView loggedInView = new LoggedInView(loggedInViewModel);
-        views.add(loggedInView, loggedInView.viewName);
-
-        application.pack();
-        application.setVisible(true);
-    }
-}
