+++ src/main/java/app/Main.java
@@ -1,17 +1,49 @@
 package app;
 
-import entity.EventFinder;
-import entity.EventFinderService;
+import data_access.FileUserDataAccessObject;
+import entity.CommonUserFactory;
+import interface_adapters.ViewManagerModel;
+import interface_adapters.logged_in.LoggedInViewModel;
+import interface_adapters.login.LoginViewModel;
+import view.LoggedInView;
+import view.LoginView;
+import view.ViewManager;
+
+import javax.swing.*;
+import java.awt.*;
+import java.io.IOException;
 
 public class Main {
-
     public static void main(String[] args) {
 
-        EventFinder ef = new EventFinderService();
+        JFrame application = new JFrame("Login");
+        application.setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
 
-        System.out.println(ef.findEvents(
-                "Your classes will be weekly on Fridays at 2:00pm. Tests will be every other Thursday."));
+        CardLayout cardLayout = new CardLayout();
 
+        JPanel views = new JPanel(cardLayout);
+        application.add(views);
+
+        ViewManagerModel viewManagerModel = new ViewManagerModel();
+        new ViewManager(views, cardLayout, viewManagerModel);
+
+        LoginViewModel loginViewModel = new LoginViewModel();
+        LoggedInViewModel loggedInViewModel = new LoggedInViewModel();
+
+        FileUserDataAccessObject userDataAccessObject;
+        try {
+            userDataAccessObject = new FileUserDataAccessObject("./users.csv", new CommonUserFactory());
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+
+        LoginView loginView = LoginUseCaseFactory.create(viewManagerModel, loginViewModel, loggedInViewModel, userDataAccessObject);
+        views.add(loginView, loginView.viewName);
+
+        LoggedInView loggedInView = new LoggedInView(loggedInViewModel);
+        views.add(loggedInView, loggedInView.viewName);
+
+        application.pack();
+        application.setVisible(true);
     }
-
 }
