+++ src/main/java/entity/CommonUserFactory.java
