+++ src/main/java/interface_adapters/ViewManagerModel.java
